transvoxel working, but with seams still

+++ Assets/VoxelMaster/Core/Voxel.cs
@@ -1,4 +1,4 @@
 public struct Voxel {
-    public sbyte density;
+    public float density;
     public byte materialIndex;
 }
